chore: make some fixes and refactor
