+++ package.json
@@ -11,8 +11,8 @@
     "_deploy-common": "npm run zip && yc serverless function version create --function-name=$YC_FUNCTION_NAME --runtime=$YC_RUNTIME --entrypoint=$YC_ENTRYPOINT --memory=$YC_MEMORY --execution-timeout=$YC_EXECUTION_TIMEOUT --source-path=function.zip --cloud-id=$YC_CLOUD_ID --folder-id=$YC_FOLDER_ID --service-account-id=$YC_SERVICE_ACCOUNT --environment BOT_TOKEN=$BOT_TOKEN,YC_FOLDER_ID=$YC_FOLDER_ID,YDB_ENDPOINT=$YDB_ENDPOINT,YDB_DATABASE=$YDB_DATABASE,GEMINI_API_KEY=$GEMINI_API_KEY,SOCKS5H=$SOCKS5H && echo \"Function deployed. Attempting to initialize database...\" && yc serverless function invoke --name=$YC_FUNCTION_NAME --data \"{\\\"setup_database\\\": true}\" && echo \"Database initialization call finished.\"",
     "deploy": "env -u YC_FUNCTION_NAME -u BOT_TOKEN -u YDB_DATABASE -u YC_ENTRYPOINT dotenv -e .env -- bash -c 'npm run _deploy-common'",
     "deploy-test": "env -u YC_FUNCTION_NAME -u BOT_TOKEN -u YDB_DATABASE -u YC_ENTRYPOINT dotenv -e .envtest -- bash -c 'npm run _deploy-common'",
-    "deploy-answering": "env -u YC_FUNCTION_NAME -u BOT_TOKEN -u YDB_DATABASE -u YC_ENTRYPOINT dotenv -e .envanswering -- bash -c 'npm run _deploy-common'",
-    "deploy-answering-test": "env -u YC_FUNCTION_NAME -u BOT_TOKEN -u YDB_DATABASE -u YC_ENTRYPOINT dotenv -e .envansweringtest -- bash -c 'npm run _deploy-common'",
+    "deploy-triggered": "env -u YC_FUNCTION_NAME -u BOT_TOKEN -u YDB_DATABASE -u YC_ENTRYPOINT dotenv -e .envtriggered -- bash -c 'npm run _deploy-common'",
+    "deploy-triggered-test": "env -u YC_FUNCTION_NAME -u BOT_TOKEN -u YDB_DATABASE -u YC_ENTRYPOINT dotenv -e .envtriggeredtest -- bash -c 'npm run _deploy-common'",
     "deploy-settings": "env -u YC_FUNCTION_NAME -u BOT_TOKEN -u YDB_DATABASE -u YC_ENTRYPOINT dotenv -e .envsettings -- bash -c 'npm run _deploy-common'",
     "deploy-settings-test": "env -u YC_FUNCTION_NAME -u BOT_TOKEN -u YDB_DATABASE -u YC_ENTRYPOINT dotenv -e .envsettingstest -- bash -c 'npm run _deploy-common'",
     "clear-cache": "echo \"Для очистки кеша выполните: unset YC_FUNCTION_NAME BOT_TOKEN YDB_DATABASE YC_ENTRYPOINT\""
